// server/enums/conditionType.ts
enum ConditionType {
  ABI = 'ABI',
  ADHD = 'ADHD',
  ASC = 'ASC',
  DLD_LANG = 'DLD_LANG',
  LD_DOWN = 'LD_DOWN',
  DYSCALCULIA = 'DYSCALCULIA',
  DYSLEXIA = 'DYSLEXIA',
  DYSPRAXIA = 'DYSPRAXIA',
  FASD = 'FASD',
  DLD_HEAR = 'DLD_HEAR',
  LD_OTHER = 'LD_OTHER',
  MENTAL_HEALTH = 'MENTAL_HEALTH',
  NEURODEGEN = 'NEURODEGEN',
  PHYSICAL_OTHER = 'PHYSICAL_OTHER',
  TOURETTES = 'TOURETTES',
  VISUAL_IMPAIR = 'VISUAL_IMPAIR',
  OTHER = 'OTHER',
  DLD_OTHER = 'DLD_OTHER',
  LEARN_DIFF_OTHER = 'LEARN_DIFF_OTHER',
  LONG_TERM_OTHER = 'LONG_TERM_OTHER',
  NEURO_OTHER = 'NEURO_OTHER',
}

export default ConditionType